requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.1
django-allauth==0.51.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-rest-auth==0.9.5
django-taggit==3.0.0
django-taggit-serializer==0.1.7
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
ecdsa==0.18.0
idna==3.3
oauthlib==3.2.0
Pillow==9.4.0
pyasn1==0.4.8
pycparser==2.21
PyJWT==1.7.1
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.11
